Switch MySQL away from REPLACE to ON DUPLICATE KEY UPDATE
Fixes #189.

While working on the locks for #125 I thought I came across a state where it would be important to preserve in-place row locks
instead of doing a delete/insert combo for a new row.
